requirements.txt
# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0

# Database
Flask-SQLAlchemy>=3.1.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0

# Task Scheduling
APScheduler>=3.10.0
schedule>=1.2.0

# AI Integration
anthropic>=0.30.0

# CV matching
PyPDF2>=3.0.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# HTTP Client
urllib3>=2.0.0

# Date/Time handling
python-dateutil>=2.8.0

# Development/Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
